deployed Lambda through API Gateway
