+++ lib/lambda-stack.mjs
@@ -5,8 +5,8 @@ import { Role, ServicePrincipal, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 import { Rule, Schedule } from 'aws-cdk-lib/aws-events';
 import { LambdaFunction } from 'aws-cdk-lib/aws-events-targets';
 import { Duration } from 'aws-cdk-lib';
-import * as apigateway from 'aws-cdk-lib/aws-apigateway';
 import * as dotenv from 'dotenv';
+import { R2RStack } from './r2r-stack.mjs';
 dotenv.config();
 
 class LambdaStack extends Stack {
@@ -20,7 +20,7 @@ class LambdaStack extends Stack {
         const r2r = new lambda.Function(this, "regionFunction", {
             runtime: lambda.Runtime.NODEJS_20_X,
             handler: "region-to-region.handler",
-            code: lambda.Code.fromAsset("resources/lambdas"),
+            code: lambda.Code.fromAsset("resources/lambdas/"),
             environment: {
                 THIS_REGION: cdk.Stack.of(this).region,
                 TABLE_NAME: table.tableName,
@@ -49,47 +49,13 @@ class LambdaStack extends Stack {
             resources: [r2r.functionArn],
         }));
 
+     
+
         // Region-to-Region: EventBridge set to  to trigger the Lambda function every hour
         new Rule(this, `schedule-${cdk.Stack.of(this).region}`, {
             schedule: Schedule.rate(Duration.hours(1)),
             targets: [new LambdaFunction(r2r)],
         });
-
-        //REGION-TO-CLIENT
-
-        // Region-to-Client:  API Gateway triggers the Lambda functions
-        const api = new apigateway.RestApi(this, 'pingApi', {
-            restApiName: 'Ping Service',
-            description: 'This service pings a given URL from various AWS regions.',
-            handler: 'backend',
-            proxy: false,
-            defaultCorsPreflightOptions: {
-                allowOrigins: ['*'],
-                allowMethods: ["GET", "POST", "OPTIONS"]
-            }
-        });
-
-        // Region-to-Client: Create a resource for region-to-client pings
-        const urlPingResource = api.root.addResource('url-ping');
-        const urlPingIntegration = new apigateway.LambdaIntegration(r2c);
-        urlPingResource.addMethod('POST', urlPingIntegration);
-        
-        //region-to-client lambda
-        const r2c = new lambda.Function(this, "urlFunction", {
-            runtime: lambda.Runtime.NODEJS_20_X,
-            handler: "region-to-client.handler",
-            code: lambda.Code.fromAsset("resources/lambdas"),
-        })
-
-        // region-to-client permissions
-        r2c.addToRolePolicy(new PolicyStatement({
-            actions: [
-                'ec2:DescribeRegions',
-            ],
-            resources: [
-                '*', 
-            ],
-        }));
 
 
     }
